+++ src/Products/PageTemplates/tests/testZopePageTemplate.py
@@ -124,11 +124,11 @@ class ZPTUtilsTests(unittest.TestCase):
 
 
 class ZPTUnicodeEncodingConflictResolution(ZopeTestCase):
-    # BBB The unicode conflict resolution feature is only available
-    # for the old Zope page template engine!
+
+    select_engine = staticmethod(useOldZopeEngine)
 
     def afterSetUp(self):
-        useOldZopeEngine()
+        self.select_engine()
         zope.component.provideAdapter(DefaultTraversable, (None,))
         zope.component.provideAdapter(HTTPCharsets, (None,))
         provideUtility(PreferredCharsetResolver,
@@ -232,6 +232,12 @@ class ZPTUnicodeEncodingConflictResolution(ZopeTestCase):
                          '<div tal:content="string:foo">foo</div>')
         self.app.REQUEST.debug.sourceAnnotations = True
         self.assertEqual(zpt.pt_render().startswith('<!--'), True)
+
+
+class ZPTUnicodeEncodingConflictResolution_chameleon(
+        ZPTUnicodeEncodingConflictResolution):
+
+    select_engine = staticmethod(useChameleonEngine)
 
 
 class ZopePageTemplateFileTests(ZopeTestCase):